src/py/requirements.txt
requests>=2.28.1
beautifulsoup4>=4.12.2
opencv-python>=4.8.0.74
pytesseract>=0.3.10
Pillow>=9.4.0
selenium>=4.11.2